+++ src/GitHub_Updater/API/Bitbucket_API.php
@@ -523,7 +523,8 @@ class Bitbucket_API extends API implements API_Interface {
 	 * @return mixed $install
 	 */
 	public function remote_install( $headers, $install ) {
-		$bitbucket_org = true;
+		$bitbucket_org                     = true;
+		$options['bitbucket_access_token'] = isset( static::$options['bitbucket_access_token'] ) ? static::$options['bitbucket_access_token'] : null;
 
 		if ( 'bitbucket.org' === $headers['host'] || empty( $headers['host'] ) ) {
 			$base            = 'https://bitbucket.org';
@@ -538,6 +539,25 @@ class Bitbucket_API extends API implements API_Interface {
 
 			if ( ! empty( $install['bitbucket_username'] ) && ! empty( $install['bitbucket_password'] ) ) {
 				$install['options'][ $install['repo'] ] = "{$install['bitbucket_username']}:{$install['bitbucket_password']}";
+			}
+
+			/*
+			* Add/Save access token if present.
+			*/
+			if ( ! empty( $install['bitbucket_access_token'] ) ) {
+				$install['options'][ $install['repo'] ] = $install['bitbucket_access_token'];
+				if ( $bitbucket_org ) {
+					$install['options']['bitbucket_access_token'] = $install['bitbucket_access_token'];
+				}
+			}
+			if ( $bitbucket_org ) {
+				$token = ! empty( $install['options']['bitbucket_access_token'] )
+				? $install['options']['bitbucket_access_token']
+				: $options['bitbucket_access_token'];
+			}
+
+			if ( ! empty( static::$options['bitbucket_access_token'] ) ) {
+				unset( $install['options']['bitbucket_access_token'] );
 			}
 		}